auto-platform/03-apirunner/03-apirunner/requirements.txt
PyYaml
requests
pytest
pytest-html
jsonpath
pyyaml-include